types: move InputObjects into the types crate
