+++ crates/sui-core/src/unit_tests/authority_tests.rs
@@ -13,7 +13,11 @@ use move_core_types::{
 };
 use narwhal_executor::ExecutionIndices;
 use rand::{prelude::StdRng, SeedableRng};
+use std::collections::BTreeMap;
+use std::fs;
+use std::{convert::TryInto, env};
 use sui_adapter::genesis;
+use sui_types::object::Data;
 use sui_types::{
     base_types::dbg_addr,
     crypto::KeyPair,
@@ -23,9 +27,6 @@ use sui_types::{
     sui_system_state::SuiSystemState,
     SUI_SYSTEM_STATE_OBJECT_ID,
 };
-
-use std::fs;
-use std::{convert::TryInto, env};
 
 pub enum TestCallArg {
     Object(ObjectID),
