+++ crates/sui-types/Cargo.toml
@@ -18,6 +18,7 @@ curve25519-dalek = { version = "3", default-features = false, features = ["serde
 serde-name = "0.2.1"
 sha3 = "0.10.1"
 thiserror = "1.0.31"
+tracing = "0.1"
 hex = "0.4.3"
 serde_bytes = "0.11.6"
 serde_json = "1.0.80"
